Create circulant.py
This file is a python program to print whether the input matrix is circulant or not.The matrix is represented row-wise,with semicolons separating rows.For example,the 3x3 identity matrix is represented as the string "100;010;001". An MxN circulant matrix is a matrix which has the form:
a_1 a_2 ....a_n
a_2 a_3 ....a_1
a_3 a_4 ....a_2
...

the matrix can be of any size.
for eg - "12;21" is circulant
"123;321" is not circulant.